# openapi/polaris-mail.yaml
openapi: 3.1.0
info:
  title: polaris-mail API
  version: 0.2.0
  description: |
    HMAC-authenticated REST surface for polaris-mail. All requests are signed with
    `X-Polaris-Sig: <hex(HMAC-SHA256(secret, canonical))>` where canonical is
    `polaris-api\n<METHOD>\n<path>\n<canonical-query>\n<X-Polaris-Ts>\n<X-Polaris-Nonce>\n<sha256-hex(body)>`.

    Webhooks delivered to subscribers (the WebhookEnvelope schema below) are signed
    `X-Polaris-Sig: <hex>` over the canonical string
    `polaris-webhook\nPOST\n<path>\n<canonical-query>\n<ts>\n<nonce>\n<sha256-hex(body)>`.
    Consumers should verify with the SDK (`@polaris-mail/sdk-node`'s `verifyWebhook`
    or `polariswebhook.VerifyWebhook` from `packages/sdk-go`).
servers:
  - url: https://polaris-mail-api.workers.dev
security:
  - HMACAuth: []
components:
  securitySchemes:
    HMACAuth:
      type: apiKey
      in: header
      name: X-Polaris-Sig
  parameters:
...
    Ts:
      name: X-Polaris-Ts
      in: header
      required: true
      schema: { type: string }
    Nonce:
      name: X-Polaris-Nonce
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
paths:
...
  /v1/admin/bootstrap:
    post:
      summary: One-time seed of the first admin key. Signed with POLARIS_SECRET_A.
      security: []
      parameters:
        - { $ref: '#/components/parameters/Ts' }
        - { $ref: '#/components/parameters/Nonce' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: seeded
        '409':
          description: already consumed
...